# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlyHub API",
        "version": "1"
    },
    "paths": {
...
        "/attributes/{id}": {
            "get": {
                "tags": [
                    "Attribute"
                ],
                "summary": "Display the specified Attribute",
                "description": "Get Attribute",
                "operationId": "App\\Http\\Controllers\\API\\AttributeAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Attribute",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Attribute"
                ],
                "summary": "Update the specified Attribute in storage",
                "description": "Update Attribute",
                "operationId": "App\\Http\\Controllers\\API\\AttributeAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Attribute",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Attribute that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Attribute"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Attribute"
                ],
                "summary": "Remove the specified Attribute from storage",
                "description": "Delete Attribute",
                "operationId": "App\\Http\\Controllers\\API\\AttributeAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Attribute",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/attribute-sets/{id}": {
            "get": {
                "tags": [
                    "AttributeSet"
                ],
                "summary": "Display the specified AttributeSet",
                "description": "Get AttributeSet",
                "operationId": "App\\Http\\Controllers\\API\\AttributeSetAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of AttributeSet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "AttributeSet"
                ],
                "summary": "Update the specified AttributeSet in storage",
                "description": "Update AttributeSet",
                "operationId": "App\\Http\\Controllers\\API\\AttributeSetAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of AttributeSet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "AttributeSet that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AttributeSet"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "AttributeSet"
                ],
                "summary": "Remove the specified AttributeSet from storage",
                "description": "Delete AttributeSet",
                "operationId": "App\\Http\\Controllers\\API\\AttributeSetAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of AttributeSet",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Display the specified Category",
                "description": "Get Category",
                "operationId": "App\\Http\\Controllers\\API\\CategoryAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update the specified Category in storage",
                "description": "Update Category",
                "operationId": "App\\Http\\Controllers\\API\\CategoryAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Category that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Remove the specified Category from storage",
                "description": "Delete Category",
                "operationId": "App\\Http\\Controllers\\API\\CategoryAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/channels/{id}": {
            "get": {
                "tags": [
                    "Channel"
                ],
                "summary": "Display the specified Channel",
                "description": "Get Channel",
                "operationId": "App\\Http\\Controllers\\API\\ChannelAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Channel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Channel"
                ],
                "summary": "Update the specified Channel in storage",
                "description": "Update Channel",
                "operationId": "App\\Http\\Controllers\\API\\ChannelAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Channel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Channel that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Channel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Channel"
                ],
                "summary": "Remove the specified Channel from storage",
                "description": "Delete Channel",
                "operationId": "App\\Http\\Controllers\\API\\ChannelAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Channel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/cities/{id}": {
            "get": {
                "tags": [
                    "City"
                ],
                "summary": "Display the specified City",
                "description": "Get City",
                "operationId": "App\\Http\\Controllers\\API\\CityAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of City",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "City"
                ],
                "summary": "Update the specified City in storage",
                "description": "Update City",
                "operationId": "App\\Http\\Controllers\\API\\CityAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of City",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "City that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/City"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "City"
                ],
                "summary": "Remove the specified City from storage",
                "description": "Delete City",
                "operationId": "App\\Http\\Controllers\\API\\CityAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of City",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/customers": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get a listing of the Customers.",
                "description": "Get all Customers",
                "operationId": "App\\Http\\Controllers\\API\\CustomerAPIController::index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Store a newly created Customer in storage",
                "description": "Store Customer",
                "operationId": "App\\Http\\Controllers\\API\\CustomerAPIController::store",
                "requestBody": {
                    "description": "Customer that should be stored",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/customers/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Display the specified Customer",
                "description": "Get Customer",
                "operationId": "App\\Http\\Controllers\\API\\CustomerAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Customer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Update the specified Customer in storage",
                "description": "Update Customer",
                "operationId": "App\\Http\\Controllers\\API\\CustomerAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Customer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Customer that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Remove the specified Customer from storage",
                "description": "Delete Customer",
                "operationId": "App\\Http\\Controllers\\API\\CustomerAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Customer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/integrations/{id}": {
            "get": {
                "tags": [
                    "Integration"
                ],
                "summary": "Display the specified Integration",
                "description": "Get Integration",
                "operationId": "App\\Http\\Controllers\\API\\IntegrationAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Integration",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Integration"
                ],
                "summary": "Update the specified Integration in storage",
                "description": "Update Integration",
                "operationId": "App\\Http\\Controllers\\API\\IntegrationAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Integration",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Integration that should be updated",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Integration"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Integration"
                ],
                "summary": "Remove the specified Integration from storage",
                "description": "Delete Integration",
                "operationId": "App\\Http\\Controllers\\API\\IntegrationAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Integration",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/products/{id}/channels/{id}": {
            "delete": {
                "tags": [
                    "ChannelProduct"
                ],
                "summary": "Remove the specified ChannelProduct from storage",
                "description": "Delete ChannelProduct",
                "operationId": "App\\Http\\Controllers\\API\\ProductChannelAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of ChannelProduct",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/products/{id}/images/{id}": {
            "delete": {
                "tags": [
                    "ProductImage"
                ],
                "summary": "Remove the specified ProductImage from storage",
                "description": "Delete ProductImage",
                "operationId": "App\\Http\\Controllers\\API\\ProductImageAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of ProductImage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AttributeSet": {
                "required": [
                    ""
                ],
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "deleted_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "tenant_id": {
                        "description": "tenant_id",
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AttributeSet"
                }
            },
...
        }
    },
    "security": [
        []
    ]
}